fix: reemplazar Decimal por float en modelo de producto

+++ services/catalog-service/app/models/product.py
@@ -22,7 +22,7 @@ class ProductBase(BaseModel):
     name: str = Field(..., min_length=1, max_length=200)
     description: str = Field(default="", max_length=2000)
     sku: str = Field(..., min_length=1, max_length=50)
-    price: Decimal = Field(..., gt=0, decimal_places=2)
+    price: float = Field(..., gt=0)
     stock: int = Field(default=0, ge=0)
     category: str = Field(default="general", max_length=100)
     is_active: bool = True
@@ -35,7 +35,7 @@ class ProductCreate(ProductBase):
 class ProductUpdate(BaseModel):
     name: Optional[str] = Field(None, min_length=1, max_length=200)
     description: Optional[str] = Field(None, max_length=2000)
-    price: Optional[Decimal] = Field(None, gt=0, decimal_places=2)
+    price: Optional[float] = Field(None, gt=0)
     stock: Optional[int] = Field(None, ge=0)
     category: Optional[str] = Field(None, max_length=100)
     is_active: Optional[bool] = None
